Refactors utils files Fixs constants file
