--- src/common/tokens.js
@@ -1,4 +1,0 @@
-export const tokenAmountInUnits = (amount, decimals = 18) => {
-  const decimalsPerToken = Math.pow(10, decimals)
-  return amount / decimalsPerToken
-}
