+++ src/index.js
@@ -4,15 +4,12 @@ import * as serviceWorker from './serviceWorker'
 import App from './App'
 import React from 'react'
 import ReactDOM from 'react-dom'
-import axios from 'axios'
 import { BrowserRouter as Router } from 'react-router-dom'
 import ReactGA from 'react-ga'
 import logdown from 'logdown'
-import { AXIOS_BASE_URL, GOOGLE_ANALYTICS_URL, LOGGER_ENABLED } from './common/constants'
+import { GOOGLE_ANALYTICS_URL, LOGGER_ENABLED } from './common/constants'
 const logger = logdown('Livepeer:Index')
 logger.state.isEnabled = LOGGER_ENABLED
-// Axios default cfg
-axios.defaults.baseURL = AXIOS_BASE_URL
 const googleAnalyticsURL = GOOGLE_ANALYTICS_URL
 // Google analytics
 if (process.env && process.env.NODE_ENV === 'production' && googleAnalyticsURL) {
